update APIs, multiSig size calculator

+++ APIP-Server/src/main/java/APIP16V1_Token/TokenByIds.java
@@ -75,7 +75,7 @@ public class TokenByIds extends HttpServlet {
         //response
         replier.setData(meetMap);
         replier.setGot(meetMap.size());
-        replier.setTotal(meetMap.size());
+        replier.setTotal((long) meetMap.size());
         esRequest.writeSuccess(dataCheckResult.getSessionKey());
     }
