+++ APIP-Server/src/main/java/APIP2V1_Blockchain/OpReturnByIds.java
@@ -72,7 +72,7 @@ public class OpReturnByIds extends HttpServlet {
 
         replier.setData(meetMap);
         replier.setGot(meetMap.size());
-        replier.setTotal(meetMap.size());
+        replier.setTotal((long) meetMap.size());
         esRequest.writeSuccess(dataCheckResult.getSessionKey());
     }
