+++ APIP-Server/src/main/java/APIP8V1_Group/GroupByIds.java
@@ -77,7 +77,7 @@ public class GroupByIds extends HttpServlet {
         //response
         replier.setData(meetMap);
         replier.setGot(meetMap.size());
-        replier.setTotal(meetMap.size());
+        replier.setTotal((long) meetMap.size());
         esRequest.writeSuccess(dataCheckResult.getSessionKey());
     }
