+++ APIP-Server/src/main/java/APIP9V1_Team/TeamMembers.java
@@ -73,7 +73,7 @@ public class TeamMembers extends HttpServlet {
         //Response
         replier.setData(dataMap);
         replier.setGot(dataMap.size());
-        replier.setTotal(dataMap.size());
+        replier.setTotal((long) dataMap.size());
         esRequest.writeSuccess(dataCheckResult.getSessionKey());
     }
